Configure the code environment as a ProServ environment
